requirements.txt
Django==2.2.1
djangorestframework==3.9.4
Pillow==6.0.0
django-webpack-loader==0.6.0
flake8==3.7.7
